Rename challenger into SOlver
